+++ spec/generators/install_generator_spec.rb
@@ -15,12 +15,21 @@ RSpec.describe PaperTrail::InstallGenerator, type: :generator do
     end
 
     it "generates a migration for creating the 'versions' table" do
+      expected_parent_class = lambda {
+        old_school = "ActiveRecord::Migration"
+        ar_version = ActiveRecord::VERSION
+        if ar_version::MAJOR >= 5
+          format("%s[%d.%d]", old_school, ar_version::MAJOR, ar_version::MINOR)
+        else
+          old_school
+        end
+      }.call
       expect(destination_root).to(
         have_structure {
           directory("db") {
             directory("migrate") {
               migration("create_versions") {
-                contains "class CreateVersions"
+                contains("class CreateVersions < " + expected_parent_class)
                 contains "def change"
                 contains "create_table :versions"
               }